--- MRBSM-SPH.vcxproj
@@ -1,51 +1,0 @@
-﻿<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-	<ItemGroup>
-		<ClCompile Include="libraries - mrbsm\ce300specific.c">
-			<CompileAs>CompileAsC</CompileAs>
-		</ClCompile>
-		<ClCompile Include="libraries - mrbsm\generalcodes.c">
-			<CompileAs>CompileAsC</CompileAs>
-		</ClCompile>
-		<ClCompile Include="main.c">
-			<CompileAs>CompileAsC</CompileAs>
-		</ClCompile>
-	</ItemGroup>
-
-	<PropertyGroup Label="Globals">
-		<ConfigurationType>Application</ConfigurationType>
-		<PlatformToolset>v142</PlatformToolset>
-		<CharacterSet>MultiByte</CharacterSet>
-		<ProjectGuid>{31a8b05f-0f11-4b54-8d86-ea3edbb35bf8}</ProjectGuid>
-	</PropertyGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
-		<OutDir>$(SolutionDir)$(Configuration)\</OutDir>
-		<TargetName>MRBSM</TargetName>
-		<Optimization>Disabled</Optimization>
-	</PropertyGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
-		<OutDir>$(SolutionDir)$(Configuration)\</OutDir>
-		<TargetName>MRBSM</TargetName>
-		<Optimization>MaxSpeed</Optimization>
-	</PropertyGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-		<OutDir>$(SolutionDir)$(Configuration)\</OutDir>
-		<TargetName>MRBSM</TargetName>
-		<Optimization>Disabled</Optimization>
-	</PropertyGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-		<OutDir>$(SolutionDir)$(Configuration)\</OutDir>
-		<TargetName>MRBSM</TargetName>
-		<Optimization>MaxSpeed</Optimization>
-	</PropertyGroup>
-
-	<Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
-	<Import Project="$(VCTargetsPath)\Microsoft.Cpp.props" />
-	<Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
-
-	<ImportGroup Label="ExtensionTargets">
-	</ImportGroup>
-</Project>
