--- main-mrbsm.h
@@ -1,12 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <math.h>
-#include <string.h>
-
-#include ".\libraries - mrbsm\ce300specific_MRBSM.h"
-
-#include <stdio.h>
-
-
-
-double** main_mrbsm();
